# internal/server/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "✨Reverse-engineered API for Gemini web app. It can be used as a genuine API key from OpenAI, Gemini, and Claude.",
        "title": "Gemini Free API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8787",
    "basePath": "/",
    "paths": {
...
        "/gemini/v1beta/models/{model}:generateContent": {
            "post": {
                "description": "Generates content using the Gemini model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gemini"
                ],
                "summary": "Generate Content (Gemini)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Generate Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.GeminiGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.GeminiGenerateResponse"
                        }
                    }
                }
            }
        },
        "/gemini/v1beta/models/{model}:streamGenerateContent": {
            "post": {
                "description": "Streams generated content using the Gemini model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gemini"
                ],
                "summary": "Stream Generate Content (Gemini)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Generate Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.GeminiGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chunked JSON response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/openai/v1/chat/completions": {
            "post": {
                "description": "Generates a completion for the chat message. Supports both standard JSON and streaming (SSE) response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "Chat Completions (OpenAI)",
                "parameters": [
                    {
                        "description": "Chat Completion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gemini-free-api_internal_modules_openai_dto.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of dto.ChatCompletionChunk JSON objects",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gemini-free-api_internal_modules_gemini_dto.Candidate": {
            "type": "object",
            "properties": {
                "content": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.Content"
                },
                "finishMessage": {
                    "type": "string"
                },
                "finishReason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.Content": {
            "type": "object",
            "properties": {
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.Part"
                    }
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.FunctionCall": {
            "type": "object",
            "properties": {
                "args": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.FunctionCallingConfig": {
            "type": "object",
            "properties": {
                "allowed_function_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "mode": {
                    "description": "\"AUTO\", \"ANY\", \"NONE\"",
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.FunctionDeclaration": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.FunctionResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "response": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.GeminiGenerateRequest": {
            "type": "object",
            "properties": {
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.Content"
                    }
                },
                "generationConfig": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.GenerationConfig"
                },
                "safety_settings": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "tool_config": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.ToolConfig"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.Tool"
                    }
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.GeminiGenerateResponse": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.Candidate"
                    }
                },
                "usageMetadata": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.UsageMetadata"
                }
            }
        },
...
        "gemini-free-api_internal_modules_gemini_dto.GenerationConfig": {
            "type": "object",
            "properties": {
                "maxOutputTokens": {
                    "type": "integer"
                },
                "temperature": {
                    "type": "number"
                },
                "topK": {
                    "type": "integer"
                },
                "topP": {
                    "type": "number"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.InlineData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.Part": {
            "type": "object",
            "properties": {
                "function_call": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.FunctionCall"
                },
                "function_response": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.FunctionResponse"
                },
                "inline_data": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.InlineData"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "gemini-free-api_internal_modules_gemini_dto.Tool": {
            "type": "object",
            "properties": {
                "function_declarations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.FunctionDeclaration"
                    }
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.ToolConfig": {
            "type": "object",
            "properties": {
                "function_calling_config": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_gemini_dto.FunctionCallingConfig"
                }
            }
        },
        "gemini-free-api_internal_modules_gemini_dto.UsageMetadata": {
            "type": "object",
            "properties": {
                "candidatesTokenCount": {
                    "type": "integer"
                },
                "promptTokenCount": {
                    "type": "integer"
                },
                "totalTokenCount": {
                    "type": "integer"
                }
            }
        },
        "gemini-free-api_internal_modules_openai_dto.ChatCompletionMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_openai_dto.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "description": "Optional OpenAI-compatible extension. Requests with the same ID reuse server-side Gemini web conversation context. If omitted, the server can also infer context from standard OpenAI messages history within the same process.",
                    "type": "string"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_openai_dto.ChatCompletionMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "reasoning": {
                    "description": "Responses-style reasoning options. reasoning.effort maps to Gemini web thinking level.",
                    "type": "object",
                    "properties": {
                        "effort": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high"
                            ]
                        }
                    }
                },
                "reasoning_effort": {
                    "description": "OpenAI-compatible reasoning effort. low/medium map to Standard; high maps to Extended.",
                    "type": "string",
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ]
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "thinking_level": {
                    "description": "Gemini web thinking level extension. Supported values: standard, extended.",
                    "type": "string",
                    "enum": [
                        "standard",
                        "extended"
                    ]
                },
                "tool_choice": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gemini-free-api_internal_modules_openai_dto.ToolDefinition"
                    }
                }
            }
        },
...
        "gemini-free-api_internal_modules_openai_dto.ToolDefinition": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/gemini-free-api_internal_modules_openai_dto.ToolFunctionDefinition"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "gemini-free-api_internal_modules_openai_dto.ToolFunctionDefinition": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "@SchemaType object",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}